requirements.txt
requests
google-auth
google-cloud-texttospeech
pygame